Structured new features into separate files. Documented some of the type helpers
