+++ src/forsyde/sdf_moc.hpp
@@ -24,6 +24,7 @@
 #include <functional>
 #include <tuple>
 
+#include "token.hpp"
 #include "sdf_process.hpp"
 #include "sdf_process_constructors.hpp"
 #include "sdf_helpers.hpp"
